refactor(lib): rename "types" to "common"
