+++ lib/src/common.rs
