+++ playback/src/discord.rs
@@ -3,8 +3,8 @@ use std::thread::sleep;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use discord_rich_presence::{DiscordIpc, DiscordIpcClient, activity};
+use termusiclib::common::const_unknown::{UNKNOWN_ARTIST, UNKNOWN_TITLE};
 use termusiclib::track::Track;
-use termusiclib::types::const_unknown::{UNKNOWN_ARTIST, UNKNOWN_TITLE};
 
 use crate::PlayerTimeUnit;
