+++ tui/src/ui/components/playlist.rs
@@ -5,6 +5,7 @@ use std::time::Duration;
 
 use anyhow::{Context as _, Result, anyhow, bail};
 use rand::seq::IndexedRandom;
+use termusiclib::common::const_unknown::{UNKNOWN_ALBUM, UNKNOWN_ARTIST};
 use termusiclib::config::SharedTuiSettings;
 use termusiclib::config::v2::server::LoopMode;
 use termusiclib::new_database::track_ops::TrackRead;
@@ -19,7 +20,6 @@ use termusiclib::player::{
 };
 use termusiclib::track::Track;
 use termusiclib::track::{DurationFmtShort, PodcastTrackData};
-use termusiclib::types::const_unknown::{UNKNOWN_ALBUM, UNKNOWN_ARTIST};
 use termusiclib::utils::{filetype_supported, get_parent_folder, is_playlist, playlist_get_vec};
 use tui_realm_stdlib::Table;
 use tuirealm::props::Borders;
